Added basic escaping of the five XML special characters
Added basic escaping of the five characters that need to be escaped in XML: ampersand, single quote, double quote, less than, greater than. If these characters are not escaped, the value of the xmlstring property can be invalid XML if only one of these characters is found in an element. The escaping could be further improved by also supporting attributes and line breaks in attributes.