Fix Swift 4 migration.
Changeset 81f506ab only migrated the iOS target to Swift 4.0;
the watchOS, tvOS, and OSX targets were still declared as Swift 3,
so they no longer compiled.

Fix this by moving the language declaration to the project level and
removing it from all the targets.  This sets everything to be Swift 4.