Add DSL-like initializers for document and element
This introduces two @resultBuilder-based builders: AEXMLDocumentBuilder and
AEXMLElementBuilder. The main difference is that the first produces an
AEXMLDocument an only accepts a singular AEXMLElement as its root. The latter
produces an array of AEXMLElements.

Both can be used in conjunction to create new hierarchies of elements easily
and quickly. Instead of creating child elements before adding them to their
parents, they can be created in code in a way that more closely resembles the
actual XML structure in the resulting document.